% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_data.R
\name{anonymize_data}
\alias{anonymize_data}
\title{anonymize_data}
\usage{
anonymize_data(data, col_names = NULL, algo = "md5", save = FALSE)
}
\arguments{
\item{data}{A data frame containing the data}

\item{col_names}{A character vector with the names of the columns to be anonymized}

\item{algo}{The algorithm to be used in case the data are going to be hashed. Default is "random". See details for options}

\item{save}{Whether a new file with the anonymized data should be saved. Default is \code{FALSE}}
}
\description{
Function for anonymizing tabular data
}
\details{
The function will take the selected columns and fill them in with either random numbers or one of the encryption algorithms.
Specifally, the different options for the \code{algo} argument are: \code{random}, \code{md5}, \code{sha1},
\code{crc32}, \code{sha256}, \code{sha512}, \code{xxhash32}, \code{xxhash64}, and \code{murmur32}.

In case the \code{algo} argument is set to random, the algorithm will generate random numbers from
a gaussian distribution with a mean of 0 and a standard deviation of 1 -- this is the default arguments in the
\code{base::rnorm} function.

In case the \code{save} argument is set to \code{TRUE}, a csv file is created with the anonymized data.
}
