% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dir.R
\name{check_dir}
\alias{check_dir}
\title{check_dir}
\usage{
check_dir(proj_name = "project")
}
\arguments{
\item{proj_name}{name of the project}
}
\description{
Function for checking whether a directory exists.
}
\details{
Function of checking whether a directory exists. If it does, the funtion will generate a warning.
Alternatively, the function will create a new directory named after the \code{proj_name} argument.
}
