% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_change.R
\name{commit_change}
\alias{commit_change}
\title{commit_change}
\usage{
commit_change(repo_obj, message)
}
\arguments{
\item{repo_obj}{An initialized repository}

\item{message}{The message of commit}
}
\value{
This function returns no values.
}
\description{
Function for commiting changes
}
\details{
Custom function for commiting changes in a directory
}
