% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_files.R
\name{commit_files}
\alias{commit_files}
\title{commit_files}
\usage{
commit_files(repo_obj, file_list, message)
}
\arguments{
\item{repo_obj}{An initialized repository}

\item{file_list}{A list of the files to be commited}

\item{message}{The message of commit}
}
\value{
This function returns no values.
}
\description{
Commit files
}
\details{
Short function for commiting changes in the directories of a project.
}
