% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_repo.R
\name{config_repo}
\alias{config_repo}
\title{config_repo}
\usage{
config_repo(repo_obj, user_name = NULL, user_email = NULL)
}
\arguments{
\item{repo_obj}{An initialized repository}

\item{user_name}{The user's name}

\item{user_email}{The user's email}
}
\value{
This function returns no values.
}
\description{
Configure repository
}
