% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_data.R
\name{decrypt_data}
\alias{decrypt_data}
\title{decrypt_data}
\usage{
decrypt_data(file_name, password = "password", nonce_file = "nonce")
}
\arguments{
\item{file_name}{The name of the file containing the encrypted data.}

\item{password}{Password to be used.}

\item{nonce_file}{The name of the nonce file.}
}
\value{
This function returns no values.
}
\description{
Function for decrypting tabular data
}
\details{
Function for decrypting tabular data.
}
