% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt_data.R
\name{encrypt_data}
\alias{encrypt_data}
\title{encrypt_data}
\usage{
encrypt_data(file_name, password = "password")
}
\arguments{
\item{file_name}{the name of the data file to be encrypted}

\item{password}{password to be used}
}
\value{
This function returns no values.
}
\description{
Function for encrypting the files included in the data directory
}
