% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_readmes.R
\name{folder_readmes}
\alias{folder_readmes}
\title{folder_readmes}
\usage{
folder_readmes(proj_dir = NULL)
}
\arguments{
\item{proj_dir}{Directory path where the new project is located}
}
\value{
This function returns no values.
}
\description{
Write the readme files for each subfolder when a project is created
}
