% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_repo.R
\name{init_repo}
\alias{init_repo}
\title{init_repo}
\usage{
init_repo(repo_path = getwd())
}
\arguments{
\item{repo_path}{the path of the repository. If none given, the current directory used as paht.}
}
\description{
Initialize repository
}
\details{
Wrapper function for initializing a repository.
}
