% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{make_project}
\usage{
make_project(proj_name = "project", add_dir = NULL)
}
\arguments{
\item{proj_name}{name of the project}

\item{add_dir}{character vector of additional directories expect the
default ones (see below). Default is set to NULL}
}
\value{
This function returns no values.
}
\description{
This function is used for creating a new project.
}
\details{
The function creates a directory named after the \code{proj.name}
object. Inside this directory, the following subdirectories are
created by default: \code{preregistration}, \code{data},
\code{analyses}, \code{plots}, \code{manuscript}. Additional
directories could be defined using the \code{add.dir} argument.
}
