% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rproject.R
\name{make_rproject}
\alias{make_rproject}
\title{make_rproject}
\usage{
make_rproject(proj_dir = NULL)
}
\arguments{
\item{proj_dir}{Directory path where the new project is located}
}
\value{
This function returns no values.
}
\description{
Short function for creating an R project and a basic R script
}
