% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_repo.R
\name{open_repo}
\alias{open_repo}
\title{open_repo}
\usage{
open_repo(repo_obj)
}
\arguments{
\item{repo_obj}{An initialized repository}
}
\value{
Summary of the repository.
}
\description{
Open repository
}
\details{
Open repository.
}
