% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prereg_create.R
\name{prereg_create}
\alias{prereg_create}
\title{prereg_create}
\usage{
prereg_create(file_name = "preregistration", template_name = "pss",
  overwrite = FALSE, edit = TRUE)
}
\arguments{
\item{file_name}{The name of the preregistration document. Default name is \code{preregistration}.}

\item{template_name}{The selected template to use. Default is set to "osf". See details for options.}

\item{overwrite}{Whether the preregistration file should be overwritten or not. Default to \code{FALSE}.}

\item{edit}{Whether the file should be edited. Default to \code{TRUE}.}
}
\value{
Only a warning is generated in case the file exists already and the \code{overwrite} argument is set
to FALSE.
}
\description{
Function for creating markdown file for a preregistration f.
}
\details{
The options for the \code{template_name} are: \code{pss}, \code{aspredicted}, \code{cos}. See the
\code{prereg} package for more details on these options. When the \code{aspredicted} or the \code{cos} options are used,
the function uses the templates of the \code{prereg} package of Frederik Aust.
}
