% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prereg_open.R
\name{prereg_open}
\alias{prereg_open}
\title{prereg_open}
\usage{
prereg_open(file_name = NULL)
}
\arguments{
\item{file_name}{The name of the preregistration document. Default name is \code{preregistration}.}
}
\description{
Function opening the preregistration document
}
\details{
The function will open the \code{.Rmd} file with the name defined by the \code{file_name} argument. In case
no such name is given, it will open the \code{.Rmd} file located in the preregistration directory.
In case no such files are present, or multiple preregistration documents are in the preregisrtation directory,
it will generate a warning
}
