% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prereg_render.R
\name{prereg_render}
\alias{prereg_render}
\title{prereg_render}
\usage{
prereg_render(file_name = NULL)
}
\arguments{
\item{file_name}{The name of the preregistration document. Default name is \code{preregistration}.}
}
\description{
Function for rendering the preregistration document
}
\details{
The function will render the \code{.Rmd} file with the name defined by the \code{file_name} argument. In case
no such name is given, it will render the \code{.Rmd} located in the preregistration directory. In case no such files
is present, or multiple preregistration documents are in the preregisrtation directory, it will
generate a warning
}
