% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_file.R
\name{read_data_file}
\alias{read_data_file}
\title{read_data_file}
\usage{
read_data_file(file_name = NULL, header = TRUE, sep = "\\t",
  quote = "'")
}
\arguments{
\item{file_name}{the name of the file, including the corresponding extension}

\item{header}{whether the data have a header. Default is set to TRUE}

\item{sep}{the seperation argument for the columns. Default is \code{"\t"}}

\item{quote}{the quote argument. Default is \code{"\'"}}
}
\value{
This function returns no values.
}
\description{
Function for reading data file
}
