% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{shiny_app}
\alias{shiny_app}
\title{Shiny app for the \code{pssr} package}
\usage{
shiny_app()
}
\description{
Lanches a Shiny app for performing the core analyses included
in \code{pssr}
}
\details{
The function can be called without any arguments (i.e., pssr::shiny_app()).
}
