% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_directories.R
\name{zip_dir}
\alias{zip_dir}
\title{zip_directories}
\usage{
zip_dir(dir_names, password = NULL)
}
\arguments{
\item{dir_names}{A character vector with all the directories to be zipped.}

\item{password}{password to be used.}
}
\value{
This function returns no values.
}
\description{
Make a zip file with all directories
}
