% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_selection.R
\name{zip_selection}
\alias{zip_selection}
\title{zip_selection}
\usage{
zip_selection(dir_path, sel_folders, password = NULL,
  git_include = FALSE)
}
\arguments{
\item{dir_path}{A character path with the project directory.}

\item{sel_folders}{A logic (TRUE or FALSE) vector inidcating which subfolders would be zipped (in the order that are created: analyses, data, manuscript, material, plots and preregistration).}

\item{password}{password to be used. Password is optional, if not entered then no password is used}

\item{git_include}{logical value that indicates if Git folder should be on the backup}
}
\value{
This function returns no values.
}
\description{
Make a zip file with a selection of the default subfolders: analyses, data, manuscript, material, plots and preregistration
}
